fix bigint value

+++ lib/src/model/uco_transfer.dart
@@ -12,8 +12,7 @@ class UCOTransfer {
   String? to;
 
   factory UCOTransfer.fromJson(Map<String, dynamic> json) => UCOTransfer(
-        amount:
-            json['amount'] == null ? null : toBigInt(json['amount'].toDouble()),
+        amount: json['amount'] == null ? null : toBigInt(json['amount']),
         to: json['to'],
       );
